ATR-635: created a base DelegateWalkerBase class with base logic for delegate expression analysis in recursive walkers
