+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/LongOperationStart/StartLongOperationDelegateWalker.cs
@@ -7,15 +7,15 @@ using System.Linq;
 using System.Threading;
 
 using Acuminator.Utilities.Common;
-using Acuminator.Utilities.Roslyn;
 using Acuminator.Utilities.Roslyn.Semantic;
+using Acuminator.Utilities.Roslyn.Walkers;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace Acuminator.Analyzers.StaticAnalysis.LongOperationStart
 {
-	public class StartLongOperationDelegateWalker : NestedInvocationWalker
+	public class StartLongOperationDelegateWalker : DelegatesWalkerBase
 	{
 		private readonly HashSet<SyntaxNode> _delegates = new HashSet<SyntaxNode>();
 		private readonly PXContext _pxContext;
@@ -52,29 +52,13 @@ namespace Acuminator.Analyzers.StaticAnalysis.LongOperationStart
 				return;
 			}
 
-			var delegateBody = GetLongRunDelegateBody(firstArgument);
+			var delegateBody = GetDelegateNode(firstArgument);
 			if (delegateBody == null)
 			{
 				return;
 			}
 
 			_delegates.Add(delegateBody);
-		}
-
-		private SyntaxNode? GetLongRunDelegateBody(ExpressionSyntax longRunDelegateExpression)
-		{
-			ThrowIfCancellationRequested();
-
-			if (longRunDelegateExpression is AnonymousFunctionExpressionSyntax anonymousFunction)
-			{
-				return anonymousFunction.Body;
-			}
-			else
-			{
-				var symbol = GetSymbol<ISymbol>(longRunDelegateExpression);
-
-				return symbol?.DeclaringSyntaxReferences.FirstOrDefault()?.GetSyntax(CancellationToken);
-			}
 		}
 	}
 }
